refactor: split sparx/element.rb into individual files
Split 453-line element.rb into 22 one-class-per-file modules under
lib/xmi/sparx/element/ with autoload. Pure file reorganization,
zero logic change.
