+++ lib/xmi/sparx/element/code.rb
@@ -1,0 +1,19 @@
+# frozen_string_literal: true
+
+module Xmi
+  module Sparx
+    module Element
+      class Code < Lutaml::Model::Serializable
+        attribute :gentype, :string
+        attribute :product_name, :string
+
+        xml do
+          root "code"
+
+          map_attribute "gentype", to: :gentype
+          map_attribute "product_name", to: :product_name
+        end
+      end
+    end
+  end
+end
